moved buildMiddie
