+++ lib/fourOhFour.js
@@ -19,6 +19,7 @@ const {
   kHooks
 } = require('./symbols.js')
 const { beforeHandlerWarning } = require('./warnings')
+const { buildMiddie } = require('./middleware')
 
 /**
  * Each fastify instance have a:
@@ -58,7 +59,7 @@ function fourOhFour (options) {
     context[kFourOhFourContext] = _404Context
   }
 
-  function setNotFoundHandler (opts, handler, avvio, routeHandler, buildMiddie) {
+  function setNotFoundHandler (opts, handler, avvio, routeHandler) {
     // First initialization of the fastify root instance
     if (this[kCanSetNotFoundHandler] === undefined) {
       this[kCanSetNotFoundHandler] = true
@@ -110,12 +111,12 @@ function fourOhFour (options) {
     }
 
     this.after((notHandledErr, done) => {
-      _setNotFoundHandler.call(this, prefix, opts, handler, avvio, routeHandler, buildMiddie)
+      _setNotFoundHandler.call(this, prefix, opts, handler, avvio, routeHandler)
       done(notHandledErr)
     })
   }
 
-  function _setNotFoundHandler (prefix, opts, handler, avvio, routeHandler, buildMiddie) {
+  function _setNotFoundHandler (prefix, opts, handler, avvio, routeHandler) {
     const context = new Context(
       opts.schema,
       handler,
