+++ lib/middleware.js
@@ -1,9 +1,10 @@
 'use strict'
 
+const Middie = require('middie')
 const handleRequest = require('./handleRequest')
 const { hookRunner, hookIterator } = require('./hooks')
 
-module.exports.onRunMiddlewares = function onRunMiddlewares (err, req, res, reply) {
+function onRunMiddlewares (err, req, res, reply) {
   if (err != null) {
     reply.send(err)
     return
@@ -20,4 +21,18 @@ module.exports.onRunMiddlewares = function onRunMiddlewares (err, req, res, repl
   } else {
     handleRequest(null, reply.request, reply)
   }
+}
+module.exports.onRunMiddlewares = onRunMiddlewares
+
+module.exports.buildMiddie = function buildMiddie (middlewares) {
+  if (!middlewares.length) {
+    return null
+  }
+
+  const middie = Middie(onRunMiddlewares)
+  for (var i = 0; i < middlewares.length; i++) {
+    middie.use.apply(middie, middlewares[i])
+  }
+
+  return middie
 }
