+++ lib/route.js
@@ -5,7 +5,7 @@
 const FindMyWay = require('find-my-way')
 const proxyAddr = require('proxy-addr')
 const Context = require('./context')
-const { onRunMiddlewares } = require('./middleware')
+const { buildMiddie, onRunMiddlewares } = require('./middleware')
 const { hookRunner, hookIterator } = require('./hooks')
 const supportedMethods = ['DELETE', 'GET', 'HEAD', 'PATCH', 'POST', 'PUT', 'OPTIONS']
 const validation = require('./validation')
@@ -34,7 +34,6 @@ module.exports = function buildRouting (options) {
   schemaCache.put = schemaCache.set
 
   let avvio
-  let buildMiddie
   let fourOhFour
   let trustProxy
   let requestIdHeader
@@ -52,7 +51,6 @@ module.exports = function buildRouting (options) {
   return {
     fill (args) {
       avvio = args.avvio
-      buildMiddie = args.buildMiddie
       fourOhFour = args.fourOhFour
       trustProxy = args.trustProxy
       requestIdHeader = args.requestIdHeader
